+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.4.0 (unreleased)
+- Rename `node.move()` -> `node.move_to()`
 
 ## 0.3.0 (2022-08-01)